modification ange

+++ src/Controller/AccountAddressController.php
@@ -87,8 +87,7 @@ class AccountAddressController extends AbstractController
     {
         $address=$this->entityManager->getRepository(Address::class)->findOneById($id);
 
-        if($address || $address->getUser()== $this->getUser()){
-            return $this->redirectToRoute('account_address');
+        if($address && $address->getUser()== $this->getUser()){
             $this->entityManager->remove($address);
             $this->entityManager->flush();
         }
